requirements.txt
# Core requirements
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Development
black==23.12.1
flake8==7.0.0
isort==5.12.0

# Add these to requirements.txt
streamlit==1.29.0
plotly==5.18.0

# Core Data Processing
#pandas==2.1.4
#numpy==1.26.2

# API and HTTP Requests
#requests==2.31.0

# Database Connectivity
#sqlalchemy>=1.4,<2.1
#psycopg2-binary==2.9.9

# Data Pipeline and Workflow Orchestration
#apache-airflow==2.6.2
#kafka-python==2.0.2

# Testing
#pytest==7.4.4

# Optional Visualization (if using Streamlit)
#streamlit==1.29.0

# Data Validation
#great-expectations==0.17.13

# Logging
#python-json-logger==2.0.7

# Environment and Development
#python-dotenv==1.0.0

